Enforced v0 for attribute 'external' (#3436)
