+++ crates/cairo-lang-starknet/cairo_level_tests/abi_dispatchers_tests.cairo
@@ -24,22 +24,22 @@ mod TestContract {
         another_as_library: IAnotherContractLibraryDispatcher
     }
 
-    #[external]
+    #[external(v0)]
     fn get_another_address(self: @ContractState) -> ContractAddress {
         self.another.read().contract_address
     }
 
-    #[external]
+    #[external(v0)]
     fn set_another_address(ref self: ContractState, contract_address: ContractAddress) {
         self.another.write(IAnotherContractDispatcher { contract_address });
     }
 
-    #[external]
+    #[external(v0)]
     fn get_another_class_hash(self: @ContractState) -> ClassHash {
         self.another_as_library.read().class_hash
     }
 
-    #[external]
+    #[external(v0)]
     fn set_another_class_hash(ref self: ContractState, class_hash: ClassHash) {
         self.another_as_library.write(IAnotherContractLibraryDispatcher { class_hash });
     }
