+++ crates/cairo-lang-starknet/cairo_level_tests/storage_access.cairo
@@ -52,12 +52,12 @@ mod TestContract {
         data: AbcEtc, 
     }
 
-    #[external]
+    #[external(v0)]
     fn set_data(ref self: ContractState, value: AbcEtc) {
         self.data.write(value);
     }
 
-    #[external]
+    #[external(v0)]
     fn get_data(self: @ContractState) -> AbcEtc {
         self.data.read()
     }
