+++ crates/cairo-lang-starknet/test_data/account.cairo
@@ -48,7 +48,7 @@ mod Account {
     }
 
 
-    #[external]
+    #[external(v0)]
     fn __validate_deploy__(
         self: @ContractState,
         class_hash: felt252,
