+++ crates/cairo-lang-starknet/test_data/hello_starknet.cairo
@@ -6,13 +6,13 @@ mod HelloStarknet {
     }
 
     // Increases the balance by the given amount.
-    #[external]
+    #[external(v0)]
     fn increase_balance(ref self: ContractState, amount: felt252) {
         self.balance.write(self.balance.read() + amount);
     }
 
     // Returns the current balance.
-    #[external]
+    #[external(v0)]
     fn get_balance(self: @ContractState) -> felt252 {
         self.balance.read()
     }
