+++ crates/cairo-lang-starknet/test_data/minimal_contract.cairo
@@ -2,6 +2,6 @@
 mod MinimalContract {
     #[storage]
     struct Storage {}
-    #[external]
+    #[external(v0)]
     fn empty(ref self: ContractState) {}
 }
