+++ crates/cairo-lang-starknet/test_data/test_contract.cairo
@@ -21,7 +21,7 @@ mod TestContract {
         1
     }
 
-    #[external]
+    #[external(v0)]
     fn test(ref self: ContractState, ref arg: felt252, arg1: felt252, arg2: felt252) -> felt252 {
         let mut x = self.my_storage_var.read();
         x += 1;
@@ -29,23 +29,23 @@ mod TestContract {
         x + internal_func()
     }
 
-    #[external]
+    #[external(v0)]
     fn another_function(ref self: ContractState, x: MyType) {}
 
-    #[external]
+    #[external(v0)]
     fn call_foo(
         ref self: ContractState, another_contract_address: starknet::ContractAddress, a: u128
     ) -> u128 {
         IAnotherContractDispatcher { contract_address: another_contract_address }.foo(a)
     }
 
-    #[external]
+    #[external(v0)]
     fn libcall_foo(ref self: ContractState, a: u128) -> u128 {
         IAnotherContractLibraryDispatcher { class_hash: starknet::class_hash_const::<0>() }.foo(a)
     }
 
     /// An external method that requires the `segment_arena` builtin.
-    #[external]
+    #[external(v0)]
     fn segment_arena_builtin(ref self: ContractState, ) {
         let x = felt252_dict_new::<felt252>();
         x.squash();
