+++ docs/reference/src/components/cairo/modules/language_constructs/pages/contracts.adoc
@@ -13,7 +13,7 @@ mod my_contract {
         m: LegacyMap::<felt252, u128>,
     }
 
-    #[external]
+    #[external(v0)]
     fn foo(value: felt252) -> felt252 {
         let y = x::read() + value;
         m::write(y, 3_u128);
@@ -60,10 +60,10 @@ When a contract is deployed, all of its storage variables are default-initialize
 Starknet contracts have no main() function. Instead, each function may be annotated as a special
 function using one of the attributes below:
 
-- `#[external]` functions may be called by Starknet users or by other contracts. For example:
+- `#[external(v0)]` functions may be called by Starknet users or by other contracts. For example:
 [source]
 ----
-    #[external]
+    #[external(v0)]
     fn external_example(ref self: ContractState, value: felt252) {
         x::write(value);
     }
@@ -73,7 +73,7 @@ The `ref` modifier means that the state may change - by writing to storage, for 
 Another possibility is to get a read-only reference to the storage:
 [source]
 ----
-    #[external]
+    #[external(v0)]
     fn view_example(self: @ContractState) -> felt252 {
         x::read(value)
     }
@@ -140,13 +140,9 @@ For example, the contract interface for `my_contract` above can be manually writ
 ----
 #[starknet::interface]
 trait IMyContract {
-    #[external]
     fn foo(value: felt252) -> felt252;
 }
 ----
-
-Note that the same `\#[external]`/`#[external]`/`#[event]` attributes must be used in the contract
-interface as in an implemented contract.
 
 === Using the Contract Interface Dispatcher
 
@@ -161,7 +157,6 @@ in the following way:
 ----
 #[starknet::interface]
 trait IMyContract<TContractState> {
-    #[external]
     fn foo(ref self: TContractState, value: felt252) -> felt252;
 }
 
@@ -170,7 +165,7 @@ mod MySecondContract {
     use super::IMyContractDispatcherTrait;
     use super::IMyContractDispatcher;
 
-    #[external]
+    #[external(v0)]
     fn call_foo(
         another_contract_address: starknet::ContractAddress,
         a: felt252
@@ -191,7 +186,6 @@ This can be done using the library-dispatcher in the following way:
 ----
 #[starknet::interface]
 trait IMyContract {
-    #[external]
     fn foo(value: felt252);
 }
 
@@ -200,7 +194,7 @@ mod MySecondContract {
     use super::IMyContractDispatcherTrait;
     use super::IMyContractLibraryDispatcher;
 
-    #[external]
+    #[external(v0)]
     fn libcall_foo(a: felt252) -> felt252 {
         IMyContractLibraryDispatcher { class_hash: starknet::class_hash_const::<0x1234>() }.foo(a)
     }
@@ -226,7 +220,7 @@ keccak hash of the function name - in this case `starknet_keccak("foo")`.
 ----
 #[starknet::contract]
 mod MySecondContract {
-    #[external]
+    #[external(v0)]
     fn syscall_call_another_contract(
         address: starknet::ContractAddress, selector: felt252, calldata: Array<felt252>
     ) -> Span::<felt252> {
